## Summary

- Add `[[nodiscard]]` to 7 value-returning functions missing the
attribute
  (SymbolKey ops, PositionManager key factories, onMarketEvent, try_pop)
- Add `[[unlikely]]` to 5 error/guard branches in RiskManager and
SimpleMarketMakingStrategy to keep trade-processing code in L1-hot
layout
- Parenthesize `(std::min)` in PositionManager.cpp for MSVC
`<windows.h>`
  macro compatibility
- Replace `size_t` with `uint32_t` in `pin_thread_to_core` signature and
  `launch_consumers` loop index
- Fix test discarding `try_pop` return value (now caught by
`[[nodiscard]]`)

Closes #35

+++ src/TradingEngine.cpp
@@ -26,7 +26,7 @@ void signal_handler(const int signum) {
 }
 } // namespace
 
-void pin_thread_to_core(std::thread &t, size_t core_id) {
+void pin_thread_to_core(std::thread &t, uint32_t core_id) {
 #if defined(__linux__) || defined(__gnu_linux__)
   cpu_set_t cpuset;
   CPU_ZERO(&cpuset);
@@ -111,7 +111,7 @@ void TradingEngine::launch_gateway() {
 }
 
 void TradingEngine::launch_consumers() {
-  for (size_t i = 0; i < symbols_.size(); ++i) {
+  for (uint32_t i = 0; i < symbols_.size(); ++i) {
     const auto &symbol = symbols_[i];
     auto callback = [this](const Order &order) {
       this->order_queue_->push(order);
